refactored extract-errors
