--- scripts/error-codes/safeParseJSON.js
@@ -1,21 +1,0 @@
-/**
- * Copyright (c) 2013-present, Facebook, Inc.
- * All rights reserved.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
- */
-'use strict';
-
-function safeParseJSON(path) {
-  var result;
-  try {
-    result = require(path);
-  } catch (e) {
-    result = {};
-  }
-  return result;
-}
-
-module.exports = safeParseJSON;
